+++ source/Knyaz.Optimus/ScriptExecuting/Jint/JintJsScriptExecutor.cs
@@ -1,33 +1,33 @@
 using System;
 using System.IO;
-using Jurassic;
+using Jint.Runtime;
 using Knyaz.Optimus.Dom;
 using Knyaz.Optimus.Dom.Elements;
 using Knyaz.Optimus.Dom.Events;
 using Knyaz.Optimus.Dom.Interfaces;
 using Knyaz.Optimus.Dom.Perf;
-using Knyaz.Optimus.ScriptExecuting;
 using Knyaz.Optimus.Tools;
 
-namespace Knyaz.Optimus.Scripting.Jurassic
+namespace Knyaz.Optimus.ScriptExecuting.Jint
 {
-    internal class ScriptExecutor : IScriptExecutor
+	/// <summary>
+	/// Facade for the script execution engine based on Jint.
+	/// </summary>
+	internal class JintJsScriptExecutor : IJsScriptExecutor
 	{
 		private readonly IWindowEx _window;
-		private readonly Func<Func<Stream, object>, XmlHttpRequest> _createXmlHttpRequest;
 
-		private IJsEngine _jsEngine;
+		private JintJsEngine _jsEngine;
 		
-		public ScriptExecutor(IWindowEx window, Func<Func<Stream, object>, XmlHttpRequest> createXmlHttpRequest)
+		public JintJsScriptExecutor(IWindowEx window, Func<Func<Stream, object>, XmlHttpRequest> createXmlHttpRequest)
 		{
 			_window = window ?? throw new ArgumentNullException(nameof(window));
-			_createXmlHttpRequest = createXmlHttpRequest;
 			CreateEngine(window, createXmlHttpRequest);
 		}
 
 		private void CreateEngine(IWindowEx window, Func<Func<Stream, object>, XmlHttpRequest> createXmlHttpRequest)
 		{
-			_jsEngine = new JurassicJsEngine();
+			_jsEngine = new JintJsEngine();
 			_jsEngine.Execute("var window = this");
 			_jsEngine.Execute("var self = window");
 			_jsEngine.AddGlobalType(typeof(Node));
@@ -47,6 +47,7 @@ namespace Knyaz.Optimus.Scripting.Jurassic
 			_jsEngine.AddGlobalType(typeof(Document));
 			_jsEngine.AddGlobalType(typeof(Text));
 			_jsEngine.AddGlobalType(typeof(Attr));
+			//Perf types
 			_jsEngine.AddGlobalType(typeof(ArrayBuffer));
 			_jsEngine.AddGlobalType(typeof(Int8Array));
 			_jsEngine.AddGlobalType(typeof(UInt8Array));
@@ -57,14 +58,13 @@ namespace Knyaz.Optimus.Scripting.Jurassic
 			_jsEngine.AddGlobalType(typeof(Float32Array));
 			_jsEngine.AddGlobalType(typeof(Float64Array));
 			_jsEngine.AddGlobalType(typeof(DataView));
-			
+
 			_jsEngine.SetGlobal(window);
-
-
-			_jsEngine.AddGlobalType(typeof(Event),"Event", new []{typeof(string),typeof(EventInitOptions)}, 
+			
+			_jsEngine.AddGlobalType<Event>(new []{typeof(string),typeof(EventInitOptions)}, 
 				args => new Event(window.Document, args[0]?.ToString(), args.Length > 1 ? (EventInitOptions)args[1] : null));
 			
-			_jsEngine.AddGlobalType(typeof(Event), "Image", new []{typeof(int), typeof(int)}, args => {
+			_jsEngine.AddGlobalType<Image>(new []{typeof(int), typeof(int)}, args => {
 				var img = (HtmlImageElement)_window.Document.CreateElement("img");
 				
 				if (args.Length > 0)
@@ -78,74 +78,34 @@ namespace Knyaz.Optimus.Scripting.Jurassic
 			
 			Func<Stream, object> parseJsonFn = s => _jsEngine.ParseJson(s.ReadToEnd());
 
-			_jsEngine.AddGlobalType("XMLHttpRequest", typeof(XmlHttpRequest), x => createXmlHttpRequest(parseJsonFn));
-			
+			_jsEngine.AddGlobalType("XMLHttpRequest", x => createXmlHttpRequest(parseJsonFn));
 		}
 
-		public void Execute(string type, string code)
+		public void Execute(string code)
 		{
 			if (code == null) //if error occurred on script loading.
 				return;
 
-			if (string.IsNullOrEmpty(type) || type.ToLowerInvariant() == "text/javascript")
-			{
-				try
-				{
-					_jsEngine.Execute(code);
-				}
-				catch (JavaScriptException e)
-				{
-					var ex = new ScriptExecutingException(e.Message, e, code);
-					if (OnException != null)
-						OnException(ex);
-				}
-				catch (Exception e)
-				{
-					if (OnException != null)
-						OnException(e);
-				}
-			}
-		}
-
-		public object Evaluate(string type, string code)
-		{
-			if (string.IsNullOrEmpty(type) || type.ToLowerInvariant() == "text/javascript")
-			{
-				try
-				{
-					return _jsEngine.Evaluate(code);
-				}
-				catch (JavaScriptException e)
-				{
-					return new ScriptExecutingException(e.Message, e, code);
-				}
-			}
-
-			throw new Exception("Unsupported script type: " + type);
-		}
-
-		public event Action<Exception> OnException;
-		public void Clear() => CreateEngine(_window, _createXmlHttpRequest);
-
-		public object EvalFuncAndCall(string code, object @this, params object[] args)
-		{
-			var funcCode = "(function(){ return "+code+";})()";
-
-			var func = Evaluate("text/javascript", funcCode) as Func<object, object[], object>;
-
 			try
 			{
-				return func(@this, args);
+				_jsEngine.Execute(code);
 			}
 			catch (JavaScriptException e)
 			{
-				OnException?.Invoke(new ScriptExecutingException(e.Message, e, code));
+				throw new ScriptExecutingException(e.Error.ToString(), e, code);
 			}
-			catch (Exception e)
+		}
+
+		public object Evaluate(string code)
+		{
+			try
 			{
-				OnException?.Invoke(e);
+				return _jsEngine.Evaluate(code);
 			}
-			return null;
+			catch (JavaScriptException e)
+			{
+				return new ScriptExecutingException(e.Error.ToString(), e, code);
+			}
 		}
 	}
 }